Logic for new product woks for me info read description
Docker-compose image corrected. Package.json docker_build to docker_build_and_push changed.
