+++ src/controllers/product.ts
@@ -1,5 +1,6 @@
 import ProductModel from "../models/product";
 import { Request, Response, NextFunction } from 'express';
+import { sendBarcodeToRabbitMQ } from "../utils/send_barcode";
 
 export const getAllProducts = (req: Request, res: Response) => {
     const products = ProductModel.find()
@@ -18,6 +19,8 @@ export const getProduct = (req: Request, res: Response) => {
             res.status(200).json({product: product})
         } else {
             res.status(204).json({product: product})
+            console.log("Product gets scraped")
+            sendBarcodeToRabbitMQ(barcode)
         }
     })
     .catch(err => console.log(err));
